fix(visual): CPUロードをbfloat16化 — fp32ロードのOOMサーバークラッシュを解消
evaluator実機(AC3)でバックエンドが視覚インデックス構築中の重みロード89%で
プロセスごと消失(トレースバック無し=ネイティブOOM即死)。fp32ロードは
2Bで常駐約8GB+チェックポイント(bf16)からの型変換ピークが加わり、サーバー
プロセスの既存使用分と合わさって落ちた。単独プロセスのスモークでは成功して
いたことと整合。

チェックポイントのネイティブdtypeであるbfloat16でロードするよう変更。
実測(単独スモーク): RSS 8-9GB→4.4GB、ロード13.3s(変換消滅で高速化)。
トレードオフ: 画像埋め込みは12.2s→52.7s/ページ(AVX512-BF16無しのCPUでは
bf16演算がエミュレーションになるため)。安定性を優先し、速度はCUDA版torch
導入(別途判断)で解決する方針。

+++ core/visual/encoder.py
@@ -48,7 +48,12 @@ class _TransformersBackend:
 
         self._torch = torch
         self._device = "cuda" if torch.cuda.is_available() else "cpu"
-        dtype = torch.float16 if self._device == "cuda" else torch.float32
+        # CPU は bfloat16(チェックポイントのネイティブdtype)でロードする。
+        # float32 だと 2B で常駐約8GB+ロード時の型変換ピークが加わり、サーバー
+        # プロセスの既存使用分と合わさって OOM 即死する(evaluator実機で確認:
+        # 重みロード89%でプロセスごと消失、トレースバック無し)。bf16 なら
+        # 常駐約4GB・変換なしで、単独スモークと同条件になる。
+        dtype = torch.float16 if self._device == "cuda" else torch.bfloat16
         log.info("visual_encoder_loading", model=model_name, device=self._device)
         self._model = SentenceTransformer(
             model_name,
